requirements.txt
Flask
Flask-PyMongo
Pillow
pytesseract
Flask-Session